+++ activiti-api-process-model/src/main/java/org/activiti/api/process/model/StartMessageDeploymentDefinition.java
@@ -5,5 +5,5 @@ public interface StartMessageDeploymentDefinition {
 
     ProcessDefinition getProcessDefinition();
 
-    MessageEventSubscription getMessageEventSubscription();
+    MessageSubscription getMessageEventSubscription();
 }
